# bigstring-intrinsics: use them